fix: support pg connection uri for migrations
