+++ src/postgres/migrations.ts
@@ -1,30 +1,50 @@
 import PgMigrate from 'node-pg-migrate';
 import { MigrationDirection } from 'node-pg-migrate/dist/types';
 import { logger } from '../logger';
+import { PgConnectionArgs, standardizedConnectionArgs } from './connection';
 
-export async function runMigrations(dir: string, direction: MigrationDirection) {
+/**
+ * Run migrations in one direction.
+ * @param dir - Migrations directory
+ * @param direction - Migration direction (`'down'` or `'up'`)
+ * @param connectionArgs - Postgres connection args
+ */
+export async function runMigrations(
+  dir: string,
+  direction: MigrationDirection,
+  connectionArgs?: PgConnectionArgs
+) {
+  const args = standardizedConnectionArgs(connectionArgs, 'migrations');
   await PgMigrate({
     dir,
     direction,
     count: Infinity,
     ignorePattern: '.*map',
-    databaseUrl: {
-      host: process.env.PGHOST,
-      port: parseInt(process.env.PGPORT ?? '5432'),
-      user: process.env.PGUSER,
-      password: process.env.PGPASSWORD,
-      database: process.env.PGDATABASE,
-    },
+    databaseUrl:
+      typeof args === 'string'
+        ? args
+        : {
+            host: args.host,
+            port: args.port ? parseInt(args.port) : undefined,
+            user: args.user,
+            password: args.password,
+            database: args.database,
+          },
     migrationsTable: 'pgmigrations',
     logger: {
-      info: msg => {},
+      info: _msg => {},
       warn: msg => logger.warn(msg),
       error: msg => logger.error(msg),
     },
   });
 }
 
-export async function cycleMigrations(dir: string) {
-  await runMigrations(dir, 'down');
-  await runMigrations(dir, 'up');
+/**
+ * Cycle migrations down and up.
+ * @param dir - Migrations directory
+ * @param connectionArgs - Postgres connection args
+ */
+export async function cycleMigrations(dir: string, connectionArgs?: PgConnectionArgs) {
+  await runMigrations(dir, 'down', connectionArgs);
+  await runMigrations(dir, 'up', connectionArgs);
 }
